+++ sdk/resourcemanager/servicelinker/armservicelinker/linker_client.go
@@ -43,7 +43,7 @@ func NewLinkerClient(credential azcore.TokenCredential, options *arm.ClientOptio
 // BeginCreateOrUpdate - Create or update Linker resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - parameters - Linker details.
@@ -70,7 +70,7 @@ func (client *LinkerClient) BeginCreateOrUpdate(ctx context.Context, resourceURI
 // CreateOrUpdate - Create or update Linker resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 func (client *LinkerClient) createOrUpdate(ctx context.Context, resourceURI string, linkerName string, parameters LinkerResource, options *LinkerClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "LinkerClient.BeginCreateOrUpdate"
@@ -105,7 +105,7 @@ func (client *LinkerClient) createOrUpdateCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -117,7 +117,7 @@ func (client *LinkerClient) createOrUpdateCreateRequest(ctx context.Context, res
 // BeginDelete - Delete a Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - options - LinkerClientBeginDeleteOptions contains the optional parameters for the LinkerClient.BeginDelete method.
@@ -142,7 +142,7 @@ func (client *LinkerClient) BeginDelete(ctx context.Context, resourceURI string,
 // Delete - Delete a Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 func (client *LinkerClient) deleteOperation(ctx context.Context, resourceURI string, linkerName string, options *LinkerClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "LinkerClient.BeginDelete"
@@ -177,7 +177,7 @@ func (client *LinkerClient) deleteCreateRequest(ctx context.Context, resourceURI
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -186,7 +186,7 @@ func (client *LinkerClient) deleteCreateRequest(ctx context.Context, resourceURI
 // Get - Returns Linker resource for a given name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - options - LinkerClientGetOptions contains the optional parameters for the LinkerClient.Get method.
@@ -225,7 +225,7 @@ func (client *LinkerClient) getCreateRequest(ctx context.Context, resourceURI st
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -243,7 +243,7 @@ func (client *LinkerClient) getHandleResponse(resp *http.Response) (LinkerClient
 // NewListPager - Returns list of Linkers which connects to the resource. which supports to config both application and target
 // service during the resource provision.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - options - LinkerClientListOptions contains the optional parameters for the LinkerClient.NewListPager method.
 func (client *LinkerClient) NewListPager(resourceURI string, options *LinkerClientListOptions) *runtime.Pager[LinkerClientListResponse] {
@@ -278,7 +278,7 @@ func (client *LinkerClient) listCreateRequest(ctx context.Context, resourceURI s
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -296,7 +296,7 @@ func (client *LinkerClient) listHandleResponse(resp *http.Response) (LinkerClien
 // ListConfigurations - list source configurations for a Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - options - LinkerClientListConfigurationsOptions contains the optional parameters for the LinkerClient.ListConfigurations
@@ -336,7 +336,7 @@ func (client *LinkerClient) listConfigurationsCreateRequest(ctx context.Context,
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -354,7 +354,7 @@ func (client *LinkerClient) listConfigurationsHandleResponse(resp *http.Response
 // BeginUpdate - Operation to update an existing Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - parameters - Linker details.
@@ -380,7 +380,7 @@ func (client *LinkerClient) BeginUpdate(ctx context.Context, resourceURI string,
 // Update - Operation to update an existing Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 func (client *LinkerClient) update(ctx context.Context, resourceURI string, linkerName string, parameters LinkerPatch, options *LinkerClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "LinkerClient.BeginUpdate"
@@ -415,7 +415,7 @@ func (client *LinkerClient) updateCreateRequest(ctx context.Context, resourceURI
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -427,7 +427,7 @@ func (client *LinkerClient) updateCreateRequest(ctx context.Context, resourceURI
 // BeginValidate - Validate a Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 //   - resourceURI - The fully qualified Azure Resource manager identifier of the resource to be connected.
 //   - linkerName - The name Linker resource.
 //   - options - LinkerClientBeginValidateOptions contains the optional parameters for the LinkerClient.BeginValidate method.
@@ -452,7 +452,7 @@ func (client *LinkerClient) BeginValidate(ctx context.Context, resourceURI strin
 // Validate - Validate a Linker.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-04-01-preview
+// Generated from API version 2024-07-01-preview
 func (client *LinkerClient) validate(ctx context.Context, resourceURI string, linkerName string, options *LinkerClientBeginValidateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "LinkerClient.BeginValidate"
@@ -487,7 +487,7 @@ func (client *LinkerClient) validateCreateRequest(ctx context.Context, resourceU
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01-preview")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil